Add remove book functionality
The functionality to remove a book from the display and the array of
books was implemented.

Some changes for this functionality were added:
-when a book object is created immediately its added to the books array.
-The function that create the book cards now take the objects from the
books array and also sets a dataset attribute to the remove book button
with its respective value, this value is the index in the books array of
the book that is being used to create a new book card.