Add the change read status functionality
The functionality to change the read status of the book was implemented,
so when the button is clicked the text, text color and image within the
button changes but also this change is applied in the books array,
specifically in the object associated with the book card where the
button was clicked.

The creation of the book card was modified, before within the read
status button was only plain text and an image but now its a span with
the status text and next to it the same image, this to make the change
of the read status text more easy to implement.

Also the addRemoveBookEvent was modified to a simpler form.